+++ src/net.cpp
@@ -2544,6 +2544,29 @@ size_t CConnman::GetNodeCount(NumConnections flags)
     return nNum;
 }
 
+CNode* CConnman::GetNodeByID(NodeId id)
+{
+    AssertLockHeld(cs_vNodes);
+
+    for(CNode* pnode : vNodes) {
+        if (id == pnode->GetId()) {
+            return pnode;
+        }
+    }
+    return nullptr;
+}
+
+bool CConnman::GetNodeStats(NodeId id, CNodeStats& stats)
+{
+    LOCK(cs_vNodes);
+    CNode* node = GetNodeByID(id);
+    if (node) {
+        node->copyStats(stats);
+        return true;
+    }
+    return false;
+}
+
 void CConnman::GetNodeStats(std::vector<CNodeStats>& vstats)
 {
     vstats.clear();
@@ -2567,11 +2590,32 @@ bool CConnman::DisconnectNode(const std::string& strNode)
 bool CConnman::DisconnectNode(NodeId id)
 {
     LOCK(cs_vNodes);
-    for(CNode* pnode : vNodes) {
-        if (id == pnode->GetId()) {
-            pnode->fDisconnect = true;
-            return true;
-        }
+    CNode* node = GetNodeByID(id);
+    if (node) {
+        node->fDisconnect = true;
+        return true;
+    }
+    return false;
+}
+
+bool CConnman::SetWhitelisted(NodeId id, bool fWhitelisted)
+{
+    LOCK(cs_vNodes);
+    CNode* node = GetNodeByID(id);
+    if (node) {
+        node->fWhitelisted = fWhitelisted;
+        return true;
+    }
+    return false;
+}
+
+bool CConnman::SetManualConnection(NodeId id, bool manual)
+{
+    LOCK(cs_vNodes);
+    CNode* node = GetNodeByID(id);
+    if (node) {
+        node->m_manual_connection = manual;
+        return true;
     }
     return false;
 }